% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.colrns.glm.R
\name{predict.colrns.glm}
\alias{predict.colrns.glm}
\title{Make predictions from a "colrns.glm" object}
\usage{
\method{predict}{colrns.glm}(object, newx, type = "response")
}
\arguments{
\item{object}{Fitted "colrns.glm" object.}

\item{newx}{Matrix of new values for x at which predictions are to be made.}

\item{type}{Type "link" gives the linear predictors. Type "response" gives the fitted probabilities. Default is "response".}
}
\value{
fitted probabilities or linear predictors
}
\description{
This function makes predictions from a "colrns.glm" object
}
\examples{
### Simulated Data
library(MASS)
set.seed(1)
n=500
p=30
Cov.x <- matrix(0.2, ncol=p, nrow=p)
gp.cov <- matrix(0.6, ncol=10, nrow=10)
diag(gp.cov) <- 1
for(i in 1:3){
  Cov.x[{10*(i-1)+1}:{10*i},{10*(i-1)+1}:{10*i}] <- gp.cov
}
x_mu = matrix(0,p,1)
x = mvrnorm(n, x_mu, Cov.x)
colnames(x)<-paste("x",1:p,sep="")
beta=rep(c(1,1,1,rep(0,7)),3)
prob=exp(x\%*\%beta)/(1+exp(x\%*\%beta))
y=rbinom(n, 1, prob)
### Fit the model
colrns.glm(x, y, lambda.min=FALSE)
### Predict
predict.colrns.glm(object=fit.colrns.glm, newx=x[1:10,])
predict.colrns.glm(object=fit.colrns.glm, newx=x[1:10,], type="link")
}
\keyword{prediction}
